refactor: move UI layer code to main.go
